added .rdlm support for CFX
