+++ .gitignore
@@ -1,2 +1,2 @@
 .idx/
-
+*.csv
